feat: introduce `twoslash-protocol` package
