+++ packages/twoslash-protocol/README.md
@@ -1,0 +1,30 @@
+<img src="https://twoslash.netlify.app/logo.svg" alt="Twoslash Logo" width="150">
+
+# twoslash-protocol
+
+[![npm version][npm-version-src]][npm-version-href]
+[![npm downloads][npm-downloads-src]][npm-downloads-href]
+[![bundle][bundle-src]][bundle-href]
+[![JSDocs][jsdocs-src]][jsdocs-href]
+[![License][license-src]][license-href]
+
+[📚 Documentation](https://twoslash.netlify.app/)
+
+The protocol for Twoslash interface, a universal layer to support different language services that describe common editor features. This package also includes some common utilities for working with the protocol.
+
+## License
+
+MIT License © [Anthony Fu](https://github.com/antfu
+
+<!-- Badges -->
+
+[npm-version-src]: https://img.shields.io/npm/v/twoslash-protocol?style=flat&colorA=161514&colorB=EAB836
+[npm-version-href]: https://npmjs.com/package/twoslash-protocol
+[npm-downloads-src]: https://img.shields.io/npm/dm/twoslash-protocol?style=flat&colorA=161514&colorB=E66041
+[npm-downloads-href]: https://npmjs.com/package/twoslash-protocol
+[bundle-src]: https://img.shields.io/bundlephobia/minzip/twoslash-protocol?style=flat&colorA=161514&colorB=45627B&label=minzip
+[bundle-href]: https://bundlephobia.com/result?p=twoslash-protocol
+[license-src]: https://img.shields.io/github/license/twoslashes/twoslash.svg?style=flat&colorA=161514&colorB=45627B
+[license-href]: https://github.com/twoslashes/twoslash/blob/main/LICENSE
+[jsdocs-src]: https://img.shields.io/badge/jsdocs-reference-161514?style=flat&colorA=161514&colorB=EAB836
+[jsdocs-href]: https://www.jsdocs.io/package/twoslash-protocol
