+++ packages/twoslash-protocol/src/types.ts
@@ -1,0 +1,1 @@
+export * from './types/index'
