+++ packages/twoslash-vue/package.json
@@ -51,6 +51,7 @@
   },
   "dependencies": {
     "@vue/language-core": "^1.8.27",
-    "twoslash": "workspace:*"
+    "twoslash": "workspace:*",
+    "twoslash-protocol": "workspace:*"
   }
 }
