+++ packages/twoslash/src/types/index.ts
@@ -1,5 +1,5 @@
 export * from './instance'
-export * from './nodes'
 export * from './options'
 export * from './returns'
 export * from './handbook-options'
+export * from 'twoslash-protocol/types'
